requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
neo4j==5.15.0
openai>=1.51.0
python-dotenv>=1.0.1
pytest>=7.0.0
sentence-transformers>=2.2.2
faiss-cpu==1.7.4